Fix an expression bug
I added a test while trying to triage the issue and I'm leaving it in because why not
