+++ src/parseStoryData.ts
@@ -411,16 +411,6 @@ const parseExpression = (expression: string): Expression | null => {
     };
   }
 
-  if (expression.includes(">")) {
-    const splitted = expression.split(">");
-
-    return {
-      type: "MORE_THAN",
-      key: splitted[0].trim(),
-      value: splitted[1].trim().replace(/'/g, "")
-    };
-  }
-
   if (expression.includes(">=")) {
     const splitted = expression.split(">=");
 
@@ -431,21 +421,31 @@ const parseExpression = (expression: string): Expression | null => {
     };
   }
 
-  if (expression.includes("<")) {
-    const splitted = expression.split("<");
-
-    return {
-      type: "LESS_THAN",
-      key: splitted[0].trim(),
-      value: splitted[1].trim().replace(/'/g, "")
-    };
-  }
-
   if (expression.includes("<=")) {
     const splitted = expression.split("<=");
 
     return {
       type: "LESS_THAN_OR_EQUALS",
+      key: splitted[0].trim(),
+      value: splitted[1].trim().replace(/'/g, "")
+    };
+  }
+
+  if (expression.includes(">")) {
+    const splitted = expression.split(">");
+
+    return {
+      type: "MORE_THAN",
+      key: splitted[0].trim(),
+      value: splitted[1].trim().replace(/'/g, "")
+    };
+  }
+
+  if (expression.includes("<")) {
+    const splitted = expression.split("<");
+
+    return {
+      type: "LESS_THAN",
       key: splitted[0].trim(),
       value: splitted[1].trim().replace(/'/g, "")
     };
